refactor: one byte formatter, and resolve_database lives with the installed-state it reads
Two audit stragglers:

- resolve_database was database-metadata logic trapped in
  core/annotate.py; four modules imported it from there. It moves to
  karyoscope/installed.py, next to the installed-state loading it
  wraps (no import cycles: installed and manifest import only
  exceptions).

- info's binary-1024 size ladder and annotate's binary _human_bytes
  rendered the same database differently from download and from
  diskspace's documented decimal convention ("everything user-facing
  is expressed in these"). Both now use diskspace.format_bytes.
  download's _format_size_gb stays, with a comment: registry entries
  declare sizes in GB and the listing keeps every row in that unit so
  columns compare.
